move functions into a new folder
