+++ piccolo/query/functions/string.py
@@ -1,0 +1,9 @@
+from .base import Function
+
+
+class Upper(Function):
+    function_name = "UPPER"
+
+
+class Lower(Function):
+    function_name = "LOWER"
